+++ packages/core-ui/src/modules/welcome/container/WelcomeOs.tsx
@@ -1,17 +1,19 @@
-import withCurrentUser from 'modules/auth/containers/withCurrentUser';
-import { IUser } from 'modules/auth/types';
-import React from 'react';
-import Welcome from '../components/Welcome';
-import * as compose from 'lodash.flowright';
-import { withProps } from '@erxes/ui/src/utils/core';
-import { graphql } from '@apollo/client/react/hoc';
-import { gql } from '@apollo/client';
-import { queries } from '@erxes/ui/src/team/graphql';
+import * as compose from "lodash.flowright";
+
 import {
   BranchesMainQueryResponse,
-  DepartmentsMainQueryResponse
-} from '@erxes/ui/src/team/types';
-import { Spinner } from '@erxes/ui/src';
+  DepartmentsMainQueryResponse,
+} from "@erxes/ui/src/team/types";
+
+import { IUser } from "modules/auth/types";
+import React from "react";
+import { Spinner } from "@erxes/ui/src";
+import Welcome from "../components/WelcomeOs";
+import { gql } from "@apollo/client";
+import { graphql } from "@apollo/client/react/hoc";
+import { queries } from "@erxes/ui/src/team/graphql";
+import withCurrentUser from "modules/auth/containers/withCurrentUser";
+import { withProps } from "@erxes/ui/src/utils/core";
 
 type Props = {
   currentUser: IUser;
@@ -45,22 +47,22 @@ class WelcomeContainer extends React.Component<FinalProps> {
 export default withProps<Props>(
   compose(
     graphql<Props>(gql(queries.branchesMain), {
-      name: 'branchListQuery',
+      name: "branchListQuery",
       options: () => ({
         variables: {
-          searchValue: '',
-          withoutUserFilter: true
-        }
-      })
+          searchValue: "",
+          withoutUserFilter: true,
+        },
+      }),
     }),
     graphql<Props>(gql(queries.departmentsMain), {
-      name: 'departmentListQuery',
+      name: "departmentListQuery",
       options: () => ({
         variables: {
-          searchValue: '',
-          withoutUserFilter: true
-        }
-      })
+          searchValue: "",
+          withoutUserFilter: true,
+        },
+      }),
     })
   )(withCurrentUser(WelcomeContainer))
 );