+++ packages/plugin-instagram-api/src/controller.ts
@@ -79,29 +79,16 @@ const init = async (app) => {
     if (data.object !== 'instagram') {
       return;
     }
+    console.log(`instagram message ${JSON.stringify(data)} `);
     for (const entry of data.entry) {
       // receive chat
       if (entry.messaging) {
-        const pageId = entry.id;
-        const integration = await models.Integrations.getIntegration({
-          $and: [
-            { facebookPageIds: { $in: pageId } },
-            { kind: INTEGRATION_KINDS.MESSENGER }
-          ]
-        });
-        if (integration) {
-          await models.Accounts.getAccount({ _id: integration.accountId });
+        try {
           const messageData = entry.messaging[0];
-
-          try {
-            await receiveMessage(models, subdomain, messageData);
-
-            return res.send('success');
-          } catch (e) {
-            return res.send('error' + e);
-          }
-        } else {
-          throw new Error('no integration found');
+          await receiveMessage(models, subdomain, messageData);
+          return res.send('success');
+        } catch (e) {
+          return res.send('error' + e);
         }
       }